feat: add subscription policy service, tests, refactor auth, guards, feature counter service and validate user service

+++ src/application/use-cases/delete-service.use-case.ts
@@ -1,4 +1,9 @@
-import { Inject, Injectable, NotFoundException } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Inject,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import {
   SERVICE_REPOSITORY,
   type ServiceRespository,
@@ -15,6 +20,9 @@ export class DeleteServiceUseCase {
   async handle(data: DeleteServiceDTO): Promise<void> {
     const service = await this.serviceRepository.findById(data.serviceId);
     if (!service) throw new NotFoundException('Service not found');
+
+    if (service.companyId !== data.companyId)
+      throw new ForbiddenException('The service belong not to this company');
 
     await this.serviceRepository.delete(data.serviceId);
   }
